+++ api/src/services/apprise-api.ts
@@ -1,8 +1,11 @@
-import { execSync } from "child_process";
+import { exec } from "child_process";
+import { promisify } from "util";
 
 import { curl_escape_all } from "../helpers/curl_escape";
 import { logs } from "../processing/logs";
 import { ProcessingItemType } from "../types";
+
+const execAsync = promisify(exec);
 
 export async function appriseApiPush(item: ProcessingItemType) {
   if (!process.env.APPRISE_API_ENDPOINT) {
@@ -22,9 +25,9 @@ export async function appriseApiPush(item: ProcessingItemType) {
     const command = `curl -d '{"body":"${message}", "title":"${pushTitle}","tag":"${process.env.APPRISE_API_TAG || "all"}"}' -H "Content-Type: application/json" ${url}`;
 
     console.log(`🕖 [APPRISE] Command: ${command}`);
-    const response = await execSync(command, { encoding: "utf-8" });
+    const { stdout } = await execAsync(command, { encoding: "utf-8" });
 
-    logs(item.id, `✅ [APPRISE] API request success:\r\n${response}`);
+    logs(item.id, `✅ [APPRISE] API request success:\r\n${stdout}`);
   } catch (e: unknown) {
     logs(item.id, `❌ [APPRISE] API request error:\r\n${(e as Error).message}`);
   }